Renamed credentials_validator directory to auth.
