+++ app/src/main/java/com/koleff/kare_android/common/auth/AuthenticationNotifier.kt
@@ -1,4 +1,4 @@
-package com.koleff.kare_android.common.credentials_validator
+package com.koleff.kare_android.common.auth
 
 import com.koleff.kare_android.ui.state.BaseState
 import kotlinx.coroutines.flow.StateFlow
