+++ app/src/main/java/com/koleff/kare_android/common/auth/CredentialsAuthenticatorImpl.kt
@@ -1,4 +1,4 @@
-package com.koleff.kare_android.common.credentials_validator
+package com.koleff.kare_android.common.auth
 
 import com.koleff.kare_android.common.preferences.Preferences
 import com.koleff.kare_android.data.model.response.base_response.KareError
