+++ app/src/main/java/com/koleff/kare_android/common/auth/CredentialsValidatorImpl.kt
@@ -1,4 +1,4 @@
-package com.koleff.kare_android.common.credentials_validator
+package com.koleff.kare_android.common.auth
 
 import androidx.core.util.PatternsCompat.EMAIL_ADDRESS
 import com.koleff.kare_android.data.model.dto.UserDto
