+++ app/src/main/java/com/koleff/kare_android/domain/usecases/RegisterUseCase.kt
@@ -1,7 +1,7 @@
 package com.koleff.kare_android.domain.usecases
 
 import android.util.Log
-import com.koleff.kare_android.common.credentials_validator.CredentialsAuthenticator
+import com.koleff.kare_android.common.auth.CredentialsAuthenticator
 import com.koleff.kare_android.data.model.dto.UserDto
 import com.koleff.kare_android.data.model.response.base_response.KareError
 import com.koleff.kare_android.domain.repository.AuthenticationRepository
