+++ app/src/main/java/com/koleff/kare_android/ui/view_model/LoginViewModel.kt
@@ -2,9 +2,9 @@ package com.koleff.kare_android.ui.view_model
 
 import android.util.Log
 import androidx.lifecycle.viewModelScope
-import com.koleff.kare_android.common.credentials_validator.AuthenticationNotifier
-import com.koleff.kare_android.common.credentials_validator.Credentials
-import com.koleff.kare_android.common.credentials_validator.CredentialsAuthenticator
+import com.koleff.kare_android.common.auth.AuthenticationNotifier
+import com.koleff.kare_android.common.auth.Credentials
+import com.koleff.kare_android.common.auth.CredentialsAuthenticator
 import com.koleff.kare_android.common.di.IoDispatcher
 import com.koleff.kare_android.common.navigation.Destination
 import com.koleff.kare_android.common.navigation.NavigationController
@@ -74,6 +74,6 @@ class LoginViewModel @Inject constructor(
     }
 
     fun navigateToWelcome() {
-        onNavigationEvent(NavigationEvent.ClearBackstackAndNavigateTo(Destination.Welcome))
+        onNavigationEvent(NavigationEvent.NavigateBack)
     }
 }
