+++ app/src/main/java/com/koleff/kare_android/ui/view_model/RegisterViewModel.kt
@@ -2,7 +2,7 @@ package com.koleff.kare_android.ui.view_model
 
 import android.util.Log
 import androidx.lifecycle.viewModelScope
-import com.koleff.kare_android.common.credentials_validator.Credentials
+import com.koleff.kare_android.common.auth.Credentials
 import com.koleff.kare_android.common.di.IoDispatcher
 import com.koleff.kare_android.common.navigation.Destination
 import com.koleff.kare_android.common.navigation.NavigationController
@@ -50,6 +50,6 @@ class RegisterViewModel @Inject constructor(
     }
 
     fun navigateToWelcome(){
-        onNavigationEvent(NavigationEvent.ClearBackstackAndNavigateTo(Destination.Welcome))
+        onNavigationEvent(NavigationEvent.NavigateBack)
     }
 }
