+++ app/src/test/java/com/koleff/kare_android/authentication/data/CredentialsDataStoreFake.kt
@@ -1,7 +1,7 @@
 package com.koleff.kare_android.authentication.data
 
-import com.koleff.kare_android.common.credentials_validator.Credentials
-import com.koleff.kare_android.common.credentials_validator.CredentialsDataStore
+import com.koleff.kare_android.common.auth.Credentials
+import com.koleff.kare_android.common.auth.CredentialsDataStore
 
 class CredentialsDataStoreFake: CredentialsDataStore {
     override fun getCredentials(): Credentials? {
